Restructuring of utilities for ease of modularization
