+++ cyphercat/__init__.py
@@ -7,4 +7,3 @@ from .models import *
 from .metrics import *
 from .load_data import *
 from .SVC_Utils import *
-from .config_utils import *
